debt - move text file model telemetry into telemetry contribution
